Ensure workflow always terminates
This is a temporary workaround around a GC3Pie issue:
The GC3Pie engine only automatically removes tasks in state TERMINATED,
but not those in state STOPPED. This can create problems, for example in
case an experiment is deleted. Then the engine would keep the task in
memory and try to update it, but the persistent representation on disk
(rows in the SQL table) has already been deleted. This would cause a
database integrity error.