Workaround for stitching bug with square wells
When the well had the same number of rows & columns and there was only one option on how to lay them out, the calc_grid_coordinates_from_positions ran into an error. Still not sure whether it will work for all cases, but I'm currently just catching the bug with equal numbers of rows & cols.