Separate model fitting and prediction for classifer tools
Models are fit only on a subset of data and prediction is subsequently
performed in batches. This makes it possible to perform classification in
situations where the entire dataset doesn't fit into memory. In addition,
it would allow parallelization of predicting labels and saving them to the
database.

Performance of inserting/updating label values is further improved by
targeting individual shards on the database worker servers directely,
which gives full SQL support, including multi-row statements.