After a server restart, reload tasks upon first hit on workflow status page.
This has the downside that you need to take action to ensure that the
workflow progresses after server restart.  However, it seems not
possible to solve the problem by just saving the task IDs to a
"session" file -- since there are many uWSGI workers around, there are
also many GC3Pie Engine instances running, and we do not know which
one is going to get the HTTP request for workflow status.  Assuming
that uWSGI is routing requests from the same web client always to the
same uWSGI worker process, this should work.

A better fix, in the long run, would be to have a single GC3Pie Engine
backend process, but that's a task that requires considerably more
development effort.