Fix problem with building containers
With Riccardo's latest fix, the environment variable setup_file was already changed in the main. The setup_file variable is not used when building containers, thus leading to an error because the main function called a non existing args. Thus this workaround ensures that Riccardos fix is only applied to the VM branch of tm_deploy