Use a tasks ``NEW`` timestamp to set the value in column `created_at`
The previous code would use the time the `create_gc3pie_store()`
function was invoked to set the `created_at` attribute for every task.
This commit changes it to use the timestamp the task's state gets set
to ``NEW``.

This behavior can still be incorrect if the task's state gets reset
(as happens in `.redo()`) but still better than the previous code...