Add options for robust alignment
This adds an option to rescale images to a user-defined percentile of the channel, thus making the alignment robust against very bright dirt particles, as they appear in some images.

This API change does not interfere with the standard workflow and it is off by default.

The multiplexing test may have to be adapted to include the 'rescale_percentile' and the 'robust_align' parameters in the workflow yaml file. I will be providing a new test case where robust alignment is used.

The new feature is available both in the API and as a command-line option.