Create TM, DB, and WWW users as "system" users.
System users do have a reserved range in the user DB, which doesn't
conflict with the one of regular users -- this is especially important
when deploying TM on existing infrastructure where some UIDs could
already be allocated.

In addition, service users like `nginx` or `postgres` should have
their home in `/var/lib` or `/var/run`, definitely not in `/home`.
(We keep the home of the `tissuemaps` user in `/home` to avoid setting
up a separate NFS share for it.)