Generalize register api for multiple plates
This changes from a hardcoded path containing plate id 1 to a system where the plate id for the path is based on the plate given as input to the function. This allows experiments with additional plates to use the register functionality.
Also, this allows users to delete a plate and add a new one, while still being able to use register.