Port identify_primary_iterative from cell profiler to jterator (#23)
This module effectively implements primary segmentation as it commonly done in the Pelkmans Lab using Cell Profiler.

One caveat is that there is a C++ dependency which must be compiled with `mex IdentifySecPropagateSubfunction.cpp`; compilation of this file will need to be added to TmDeploy.

`dijkstraCP.m` and `IdentifySecPropagateSubfunction.cpp` are part of CellProfiler-1.0 and should be accompanied by the licence https://github.com/CellProfiler/CellProfiler-1.0/blob/master/LICENSE

`rplabel.m` seems to be internal Pelkmans Lab code, but authorship is unclear. Authorship is listed in the remaining files.