Change logging format: move message origin to end of line.
With the provided formatting facilities for Python loggers, it is
impossible to make a fixed-width string combining many fields.
In other words, a formatting directive like:

        %(name)32s@%(filename)-32s:%(lineno)-4d

results in difficult-to-read output like this:

        tmserver.extensions.gc3pie@gc3pie.py                       :109

thereby cancelling all readability advantages of fixed-width
tabulation...

So, move the "origin" part of the logging to the end of the line, and
we'll rely on grep output colorization or other tools to visually
extract that information.