TmDeploy: launch the `jobdaemon` process with `&`
The issue: when there are already many jobs in the jobdaemon's
session, the jobdaemon process itself takes a long time to start and
daemonize (it loads the entire session first).  So systemd (or any
other launch supervisor, probably) enters a cycle of start -> timeout
-> kill -> start again.

If the launch script uses `&` then the job daemon is started in the
background and `uwsgi` can start almost immediately.

This also means that TM will not be able to submit or manage jobs
immediately after restart (possibly for a few minutes), but the
problem will solve itself *eventually* without human intervention.

It goes without saying that this is just a cheap workaround; *the*
solution to this is rather to split the jobdaemon into its own service
file and use systemd dependencies.