Drop NginX and use uWSGI for all HTTP serving (incl. static content)
This is part of an effort to simplify the TM stack to ease debugging
and possibly have TM running in Docker containers again.

(It makes total sense to have NginX as a proxy/load-balancer in front
of a big production deployment of TM; however, the rest of the TM
software stack just does not have the stability that makes this worth.
NginX will be back with a role that integrates with the OS's way of
creating virtual hosts when such big deployments make sense and don't
crush under their own weight.)