Drop support for HTTPS.
HTTPS requires a certificate, and certificates require
publicly-accessible hosts.  The bulk of TM installations now is done
on private IP addressing space; we can add SSL back when we have a
more streamlined installation procedure.  For the moment, manually
installing `certbot` will do for the few cases where we might need to
open up a TM instance to the public.