Reimplement `assert_type` in a sane way.
The old code would walk through the decorated function's signature and
import or re-create each class for validation, which trades a
dubious convenience at code writing time for a sure performance
penalty each time the decorated function is called.

The new code does all of this when the function is decorated; this,
however, requires that all types to be checked are defined at the
moment the `assert_type()` decorator is called.  Hence, some code
rearrangement is needed, and a new function `add_assert_type` needs to
be introduced for when a method of a class needs to check a parameter
against the class being defined.