Updates to morphology (#166)
New way of computing morphology features.

This commit changes the calculation of morphology features to use the `skimage.measure.regionprops` function rather than a combination of home-made and `mahotas`-based features. This also then extends the number of morphology features that are calculated to better capture object morphology. This will be useful for more fully describing the phenotypes of single cells. The reason to replace some of the existing feature calculations with the `regionprops` function was to be consistent with the library used to calculate the new features and also to possibly avoid the NaN values that were sometimes returned in the previous implementation. I have not seen any NaN values so far for elongation and circularity so far in my testing however they may still occur for some objects.

One missing feature from both the original implementation and the regionprops function is "Roundness". This has been included from the `mahotas` library.

It should be noted that several Jterator modules such as `filter.py` and `separate_clumps.py` depend on the extract method that is modified here

For more details and a comparison with the old implementation, see:
https://github.com/TissueMAPS/TissueMAPS/pull/166#issue-262635627