Non-consecutive labels fix (#168)
This commit contains checks to protect against an issue that
came up for a `label_image` with non-consecutive object
labels. Firstly, this situation should not occur, and is due to errors
in an upstream module (a combination of `separate_clumps` generating a
1-pixel object, and `segment_secondary` removing that object during
seed expansion). These issues still need addressing separately.

This situation caused a problem in `measure_morphology` because the
index for `skimage.measure.regionprops` went out of range. Note that
the missing object caused all feature values calculated afterwards to
be assigned to the wrong label.

To combat this, I added an error that should catch these situations
during `__init__`. While there, I also noticed a bug in
https://github.com/TissueMAPS/TissueMAPS/compare/master...scottberry:morphology-fix?expand=1#diff-f77948de8e3abf8f00ab95390d03ad33L137
which does not necessarily catch all multi-labeling issues. The old
version only throws an error when a reference object contains a
subobject label greater than `ref_label`. The new version should catch
any situation when a subobject label is different from `ref_label`.

Finally, I modified the indexing for `regionprops` using an enumerate
iterator, and added a final check that the actual label measured for
that object matches the current query label. This should never
actually be the case if labels are consecutive but is useful as a
sanity check. It is crucial that feature values are never assigned to
the wrong objects.