Improved separate clumps module (#170)
Made the cutting line thinner, fixed bug in nuclei separation, added an option to remove very small objects created through the cutting and cleaned up the module.

The cutting line was shifted by 1 pixel in x & y from where it should have been. This lead to the module sometimes creating additional, tiny objects.

This was also the reason that a line dilation was necessary to ensure separation of the objects. After fixing the line shifting bug, removing the line dilation doesn't interfere with nuclei being separated, but makes the separation line narrower.

Still, the module can sometimes create tiny objects through a cut and would then not perform the cut. Instead, the new default is to remove the tiny object (if it's smaller than 10 pixels in area).

Also, there were various lines not being used anymore and the output of the module was called a mask even though it was a label image. These things were fixed here as well.

As there were this many changes and a new input was added to the module, the module version number is increased from 0.2.2 to 0.30. If one updates an existing instance, the value in the handles file needs to be corrected (or the module removed from the pipeline and added again).