Fix "TypeError: set([1, 10, 5]) is not JSON serializable"
The issue here stems from the fact that not all Python types can be
directly mapped to JSON; in particular, JSON cannot represent `set`
objects.  However, TM uses a handle's `.to_dict()` method to convert
the handle to a dictionary and then feeds this dictionary directly
into the JSON serializer.  So, try to ensure that return value of
`.to_dict()` is JSON-serializable; conversion to native Python types
happens in the handle's own constructor.