Raise memory limit for "init" and "collect" jobs to 2500MB.
Apparently, 1750MB were not enough for Micha Mueller's experiment, and
crashed the `illuminati init` task.

The whole logic of this part is flawed; memory allocation should not
be hard-coded but depend on experiment and batch size, and on the
program being run.  And, obviously, users should have a way to
override the systems' choices.